requirements/base.txt
mlflow>=2.9.0
boto3>=1.34.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
pandas>=2.1.0
